+++ src/Components/Layouts/MainLayout/mainLayout.scss
@@ -1,0 +1,3 @@
+.mainLayout{
+    background-color: #fefefe;
+}
